rename ZUIAvatar for ZUIPersonAvatar
