+++ src/zui/ZUIPersonAvatar/index.stories.tsx
@@ -1,0 +1,18 @@
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+
+import ZUIPersonAvatar from '.';
+
+export default {
+  component: ZUIPersonAvatar,
+  title: 'Atoms/ZUIAvatar',
+} as ComponentMeta<typeof ZUIPersonAvatar>;
+
+const Template: ComponentStory<typeof ZUIPersonAvatar> = (args) => {
+  return <ZUIPersonAvatar orgId={args.orgId} personId={args.personId} />;
+};
+
+export const basic = Template.bind({});
+basic.args = {
+  orgId: 1,
+  personId: 1,
+};
